// soa-lab-2-front/src/app/types/properties.ts
export enum Properties {
    ID = 'id',
    NAME = 'name',
    COORDINATE_X = 'coordinate.x',
    COORDINATE_Y = 'coordinate.y',
    CREATION_DATE = 'creationDate',
    NUMBER_OF_PARTICIPANTS = 'numberOfParticipants',
    DESCRIPTION = 'description',
    PERSON_NAME = 'person.name',
    GENRE = 'genre'
}